Improve keygen
Make the keygen a crate that can be implemented and extended from other places
(like a forc-plugin). Upgraded its dependencies to avoid depending on protoc
for building
