+++ bin/keygen/Cargo.toml
@@ -13,5 +13,5 @@ description = "Command line utilities for fuel-core key management"
 anyhow = { workspace = true }
 clap = { workspace = true, features = ["derive", "env"] }
 fuel-core-types = { workspace = true, features = ["serde", "random"] }
-libp2p-core = { version = "0.38", features = ["secp256k1"] }
+libp2p-identity = { version = "0.2.4", features = ["secp256k1", "peerid"] }
 serde_json = { workspace = true, features = ["raw_value"] }
