+++ bin/keygen/src/lib.rs
@@ -1,0 +1,6 @@
+//! Keygen crate
+
+pub const BLOCK_PRODUCTION: &str = "block_production";
+pub const P2P: &str = "p2p";
+
+pub mod keygen;
